remove build folder if exist

+++ _EmptyProject/build.sh
@@ -1,4 +1,8 @@
-rm -r build
+if [ -d build ] 
+then
+    rm -r build
+fi
+
 mkdir build
 g++ -std=c++11 app.cpp -pthread -o build/app 
 build/app
